Sync audio with Screen

+++ include/new_exe/imguiManager.h
@@ -29,7 +29,7 @@ namespace NesEmulatorGL
 
         Format m_currentFormat = Format::ORIGINAL;
 
-        std::array<bool, Format::COUNT> m_changeFormats;
+        std::array<bool, (size_t)Format::COUNT> m_changeFormats;
         
         inline static constexpr unsigned MAX_SAVE_STATES = 4;
         std::array<bool, MAX_SAVE_STATES> m_requestSaveState;
